Move ExecuteCommand into requests + request rustfmt for rustfmt tests
