+++ src/actions/mod.rs
@@ -54,7 +54,6 @@ macro_rules! parse_file_path {
 mod post_build;
 pub mod requests;
 pub mod notifications;
-mod execute_command;
 
 /// Persistent context shared across all requests and notifications.
 pub enum ActionContext {
